tree structure update
