+++ intent_client/src/constants/types.ts
@@ -1,5 +1,4 @@
 
-
 export interface NodeType {
     id: string,
     label: string,
