+++ intent_client/src/features/topology/components/ListNodeSidebar/ListNode.tsx
@@ -1,0 +1,30 @@
+import {Col, ListGroup, Row} from "react-bootstrap";
+import React from "react";
+import NodeList from "../NodeList/NodeList.tsx";
+
+
+function ListNode({nodeList, resetNodeSelection}: NodeListProps) {
+
+    return (
+        <Row>
+            <Col>
+                <h3 className={'text-center'}>Node list</h3>
+                <ListGroup variant={'flush'}>
+                    {
+                        nodeList.map(
+                            (node) => {
+                                return (
+                                    <ListGroup.Item key={node} className='bg-light text-center my-3'>
+                                        <NodeList title={node} />
+                                    </ListGroup.Item>
+                                )
+                            }
+                        )
+                    }
+                </ListGroup>
+            </Col>
+        </Row>
+    )
+}
+
+export default ListNode
