+++ intent_client/src/features/topology/components/ListNodeSidebar/ListNode.types.tsx
@@ -1,0 +1,4 @@
+interface NodeListProps {
+    nodeList: string[],
+    resetNodeSelection: () => void
+}
