+++ intent_client/src/features/topology/components/NodeList/NodeList.types.tsx
@@ -1,0 +1,3 @@
+interface ListNodeProps {
+    title: string,
+}
