--- intent_client/src/helper.ts
@@ -1,93 +1,0 @@
-import routerSvg from './assets/router.png'
-import switchSvg from './assets/switch.png'
-import serverSvg from './assets/server.jpg'
-import { Position } from '@xyflow/react';
-
-function getNodeSvg(type: string) {
-    if ( type === 'router' )
-        return  routerSvg
-    else if ( type === 'switch' )
-        return switchSvg
-    else if ( type === 'server' )
-        return serverSvg
-    else
-        return ''
-}
-
-function isObjectEmpty(obj: object) {
-    for (const prop in obj) {
-        if (Object.prototype.hasOwnProperty.call(obj, prop)) {
-            return false;
-        }
-    }
-
-    return true
-}
-
-function getNodeIntersection(intersectionNode, targetNode) {
-    // https://math.stackexchange.com/questions/1724792/an-algorithm-for-finding-the-intersection-point-between-a-center-of-vision-and-a
-    const { width: intersectionNodeWidth, height: intersectionNodeHeight } =
-        intersectionNode.measured;
-    const intersectionNodePosition = intersectionNode.internals.positionAbsolute;
-    const targetPosition = targetNode.internals.positionAbsolute;
-
-    const w = intersectionNodeWidth / 2;
-    const h = intersectionNodeHeight / 2;
-
-    const x2 = intersectionNodePosition.x + w;
-    const y2 = intersectionNodePosition.y + h;
-    const x1 = targetPosition.x + targetNode.measured.width / 2;
-    const y1 = targetPosition.y + targetNode.measured.height / 2;
-
-    const xx1 = (x1 - x2) / (2 * w) - (y1 - y2) / (2 * h);
-    const yy1 = (x1 - x2) / (2 * w) + (y1 - y2) / (2 * h);
-    const a = 1 / (Math.abs(xx1) + Math.abs(yy1) || 1);
-    const xx3 = a * xx1;
-    const yy3 = a * yy1;
-    const x = w * (xx3 + yy3) + x2;
-    const y = h * (-xx3 + yy3) + y2;
-
-    return { x, y };
-}
-
-function getEdgePosition(node, intersectionPoint) {
-    const n = { ...node.measured.positionAbsolute, ...node };
-    const nx = Math.round(n.x);
-    const ny = Math.round(n.y);
-    const px = Math.round(intersectionPoint.x);
-    const py = Math.round(intersectionPoint.y);
-
-    if (px <= nx + 1) {
-        return Position.Left;
-    }
-    if (px >= nx + n.measured.width - 1) {
-        return Position.Right;
-    }
-    if (py <= ny + 1) {
-        return Position.Top;
-    }
-    if (py >= n.y + n.measured.height - 1) {
-        return Position.Bottom;
-    }
-
-    return Position.Top;
-}
-
-function getEdgeParams(source, target) {
-    const sourceIntersectionPoint = getNodeIntersection(source, target);
-    const targetIntersectionPoint = getNodeIntersection(target, source);
-
-    const sourcePos = getEdgePosition(source, sourceIntersectionPoint);
-    const targetPos = getEdgePosition(target, targetIntersectionPoint);
-
-    return {
-        sx: sourceIntersectionPoint.x,
-        sy: sourceIntersectionPoint.y,
-        tx: targetIntersectionPoint.x,
-        ty: targetIntersectionPoint.y,
-        sourcePos,
-        targetPos,
-    };
-}
-
-export {getNodeSvg, isObjectEmpty, getEdgeParams}
