+++ intent_client/src/utils/edge.ts
@@ -1,0 +1,44 @@
+import {Position} from "@xyflow/react";
+import {getNodeIntersection} from "./node.ts";
+
+function getEdgePosition(node, intersectionPoint) {
+    const n = { ...node.measured.positionAbsolute, ...node };
+    const nx = Math.round(n.x);
+    const ny = Math.round(n.y);
+    const px = Math.round(intersectionPoint.x);
+    const py = Math.round(intersectionPoint.y);
+
+    if (px <= nx + 1) {
+        return Position.Left;
+    }
+    if (px >= nx + n.measured.width - 1) {
+        return Position.Right;
+    }
+    if (py <= ny + 1) {
+        return Position.Top;
+    }
+    if (py >= n.y + n.measured.height - 1) {
+        return Position.Bottom;
+    }
+
+    return Position.Top;
+}
+
+function getEdgeParams(source, target) {
+    const sourceIntersectionPoint = getNodeIntersection(source, target);
+    const targetIntersectionPoint = getNodeIntersection(target, source);
+
+    const sourcePos = getEdgePosition(source, sourceIntersectionPoint);
+    const targetPos = getEdgePosition(target, targetIntersectionPoint);
+
+    return {
+        sx: sourceIntersectionPoint.x,
+        sy: sourceIntersectionPoint.y,
+        tx: targetIntersectionPoint.x,
+        ty: targetIntersectionPoint.y,
+        sourcePos,
+        targetPos,
+    };
+}
+
+export {getEdgePosition, getEdgeParams}
