+++ intent_client/src/utils/helper.ts
@@ -1,0 +1,12 @@
+
+function isObjectEmpty(obj: object) {
+    for (const prop in obj) {
+        if (Object.prototype.hasOwnProperty.call(obj, prop)) {
+            return false;
+        }
+    }
+
+    return true
+}
+
+export {isObjectEmpty}
