+++ intent_client/src/utils/node.ts
@@ -1,0 +1,42 @@
+import routerSvg from "../assets/router.png";
+import switchSvg from "../assets/switch.png";
+import serverSvg from "../assets/server.jpg";
+
+function getNodeSvg(type: string) {
+    if ( type === 'router' )
+        return  routerSvg
+    else if ( type === 'switch' )
+        return switchSvg
+    else if ( type === 'server' )
+        return serverSvg
+    else
+        return ''
+}
+
+function getNodeIntersection(intersectionNode, targetNode) {
+    // https://math.stackexchange.com/questions/1724792/an-algorithm-for-finding-the-intersection-point-between-a-center-of-vision-and-a
+    const { width: intersectionNodeWidth, height: intersectionNodeHeight } =
+        intersectionNode.measured;
+    const intersectionNodePosition = intersectionNode.internals.positionAbsolute;
+    const targetPosition = targetNode.internals.positionAbsolute;
+
+    const w = intersectionNodeWidth / 2;
+    const h = intersectionNodeHeight / 2;
+
+    const x2 = intersectionNodePosition.x + w;
+    const y2 = intersectionNodePosition.y + h;
+    const x1 = targetPosition.x + targetNode.measured.width / 2;
+    const y1 = targetPosition.y + targetNode.measured.height / 2;
+
+    const xx1 = (x1 - x2) / (2 * w) - (y1 - y2) / (2 * h);
+    const yy1 = (x1 - x2) / (2 * w) + (y1 - y2) / (2 * h);
+    const a = 1 / (Math.abs(xx1) + Math.abs(yy1) || 1);
+    const xx3 = a * xx1;
+    const yy3 = a * yy1;
+    const x = w * (xx3 + yy3) + x2;
+    const y = h * (-xx3 + yy3) + y2;
+
+    return { x, y };
+}
+
+export {getNodeSvg, getNodeIntersection}
